backend/requirements.txt
django>=5.1.0
djangorestframework>=3.15.0
django-cors-headers>=4.7.0
asgiref>=3.8.0
requests>=2.31.0
sqlparse>=0.5.0 